feat(catalog): unit-conversion-normalized quantity in product pricing
Ports the shared unit-code canonicalization (CatalogUnitCodes: trim/lower +
qty->pc alias) and the upstream normalizedQuantityForPricing: when
?quantityUnit differs from a product's canonical default_unit and the product
has an active conversion for that unit, the pricing-context quantity is scaled
by the conversion factor for that product only. PricingContext.Quantity is now
decimal so a fractional normalized quantity compares correctly to the int
min/max windows.

1 HTTP test (1 box = 12.5 kg reaching a min-10 bulk tier); 341 .NET tests green.

+++ packages/dotnet/src/OpenMercato.Modules.Catalog/Lib/CatalogPricing.cs
@@ -10,10 +10,11 @@ namespace OpenMercato.Modules.Catalog.Lib;
 /// </summary>
 public static class CatalogPricing
 {
-    /// <summary>Resolution context (upstream <c>PricingContext</c>).</summary>
+    /// <summary>Resolution context (upstream <c>PricingContext</c>). Quantity is decimal so a
+    /// unit-conversion-normalized (fractional) quantity compares correctly to the int min/max windows.</summary>
     public sealed record PricingContext(
         Guid? ChannelId, Guid? OfferId, Guid? UserId, Guid? UserGroupId,
-        Guid? CustomerId, Guid? CustomerGroupId, int Quantity, DateTimeOffset Date);
+        Guid? CustomerId, Guid? CustomerGroupId, decimal Quantity, DateTimeOffset Date);
 
     /// <summary>A flattened candidate price row with the fields resolution needs (price + its price-kind
     /// code/promotion flag + the linked offer's channel).</summary>
